requirements.txt
beautifulsoup4==4.12.3
keyring==25.4.1
nodejs==0.1.1
pycryptodome==3.20.0
pycryptodomex==3.20.0

